chore: restructure helpers and tests
