+++ tests/scripts/update-managed-objects.test.js
@@ -4,13 +4,13 @@ describe('update-managed-objects', () => {
   jest.mock('fs')
   const fs = require('fs')
   const path = require('path')
-  jest.mock('../helpers/get-access-token')
-  const getAccessToken = require('../helpers/get-access-token')
+  jest.mock('../../helpers/get-access-token')
+  const getAccessToken = require('../../helpers/get-access-token')
   jest.spyOn(console, 'log').mockImplementation(() => {})
   jest.spyOn(console, 'error').mockImplementation(() => {})
   jest.spyOn(process, 'exit').mockImplementation(() => {})
 
-  const updateManagedObject = require('./update-managed-objects')
+  const updateManagedObject = require('../../scripts/update-managed-objects/update-managed-objects')
 
   const mockValues = {
     fidcUrl: 'https://fidc-test.forgerock.com',
